Staging

+++ apps/api/src/providers/storage.provider.ts
@@ -55,7 +55,9 @@ export class StorageProvider {
     const stream = res.Body as Readable;
     const chunks: Buffer[] = [];
     for await (const chunk of stream) {
-      chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
+      chunks.push(
+        Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk as Uint8Array),
+      );
     }
     return Buffer.concat(chunks);
   }
